I don't even know I'm sorry I'm such a bad octokitten
I started to make a management page for the owner of the site, but then
I realised all of my assets aRE SERVED TOGETHER WHICH RUINED LIKE
EVERYTHING.

So I disabled that and I’m probably going to put this entire day into
moving the right styles to the corresponding stylesheets of each
controller.

I don’t understand how I came this far without realising this behaviour
of Rails. Whatever! Live and learn. :D